+++ packages/components/src/components/display/display-utils.ts
@@ -1,4 +1,4 @@
-import type { BreakpointCustomizable, TextAlign } from '../../types';
+import type { BreakpointCustomizable, TypographyAlign, TypographyAlignDeprecated } from '../../types';
 import { hasSpecificSlottedTag } from '../../utils';
 
 export const DISPLAY_TAGS = ['h1', 'h2', 'h3', 'h4', 'h5', 'h6'] as const;
@@ -10,7 +10,9 @@ export type DisplaySize = (typeof DISPLAY_SIZES)[number];
 export const DISPLAY_COLORS = ['primary', 'inherit'] as const;
 export type DisplayColor = (typeof DISPLAY_COLORS)[number];
 
-export type DisplayAlign = TextAlign;
+/** @deprecated */
+export type DisplayAlignDeprecated = TypographyAlignDeprecated;
+export type DisplayAlign = TypographyAlign;
 
 export const displaySizeToTagMap: { [key in DisplaySize]: string } = {
   small: 'h3',