+++ packages/components/src/components/heading/heading-utils.ts
@@ -1,7 +1,7 @@
 import type { HeadingTag } from './heading-tag';
-import type { BreakpointCustomizable, TextAlign } from '../../types';
-import { hasSpecificSlottedTag } from '../../utils';
 import { HEADING_TAGS } from './heading-tag';
+import type { BreakpointCustomizable, TypographyAlign, TypographyAlignDeprecated } from '../../types';
+import { hasSpecificSlottedTag } from '../../utils';
 
 export const HEADING_SIZES = ['small', 'medium', 'large', 'x-large', 'xx-large', 'inherit'] as const;
 export type HeadingSize = (typeof HEADING_SIZES)[number];
@@ -9,7 +9,9 @@ export type HeadingSize = (typeof HEADING_SIZES)[number];
 export const HEADING_COLORS = ['primary', 'inherit'] as const;
 export type HeadingColor = (typeof HEADING_COLORS)[number];
 
-export type HeadingAlign = TextAlign;
+/** @deprecated */
+export type HeadingAlignDeprecated = TypographyAlignDeprecated;
+export type HeadingAlign = TypographyAlign;
 
 const headingSizeToTagMap: Record<HeadingSize, string> = {
   small: 'h6',