+++ packages/components/src/components/link-tile-model-signature/link-tile-model-signature-styles.ts
@@ -43,10 +43,12 @@ export const getComponentCss = (
       padding: `${spacingFluidLarge} ${spacingFluidMedium} ${spacingFluidMedium}`,
       ...gradientToTopStyle,
     },
-    model: {
+    signature: {
       position: 'absolute',
       top: spacingFluidMedium,
       left: spacingFluidMedium,
+      right: spacingFluidMedium,
+      display: 'flex',
     },
     heading: {
       margin: 0,