+++ packages/components/src/components/text-list/text-list-item/text-list-item-styles.ts
@@ -1,22 +1,33 @@
 import { getCss } from '../../../utils';
 import { addImportantToEachRule, hostHiddenStyles } from '../../../styles';
 import { spacingStaticMedium } from '@porsche-design-system/utilities-v2';
-import { cssVariablePseudoSuffix } from '../text-list/text-list-styles';
+import {
+  cssVariableUnorderedPseudoContent,
+  cssVariableOrderedGridColumn,
+  cssVariablePseudoSpace,
+  cssVariableOrderedPseudoSuffix,
+  cssVariableUnorderedGridColumn,
+} from '../text-list/text-list-styles';
 
 export const getComponentCss = (): string => {
   return getCss({
     '@global': addImportantToEachRule({
       ':host': {
-        display: 'list-item',
-        position: 'relative', // needed by before pseudo-element, used for ordered list
+        display: 'grid',
+        gridTemplateColumns: `var(${cssVariablePseudoSpace}) 1fr`,
+        columnGap: spacingStaticMedium,
         font: 'inherit', // ensures style can't be overwritten from outside
         color: 'inherit', // ensures style can't be overwritten from outside
-        listStyleType: 'inherit', // ensures style can't be overwritten from outside
-        paddingLeft: spacingStaticMedium, // space between ::marker/::before and list item
         ...hostHiddenStyles,
       },
       '::slotted(*)': {
-        [cssVariablePseudoSuffix]: '""', // don't show suffix "." for nested ordered list
+        [cssVariableUnorderedGridColumn]: '.625rem', // reserves space for ::before (nested unordered list)
+        [cssVariableUnorderedPseudoContent]: '"–"', // custom ::before char "–" (nested unordered list)
+        [cssVariableOrderedGridColumn]: '2rem', // reserves space for ::before (nested ordered list)
+        [cssVariableOrderedPseudoSuffix]: '""', // don't show ::before suffix "." (nested ordered list)
+      },
+      '::slotted(*:last-child)': {
+        gridColumn: 2,
       },
     }),
   });