+++ packages/components/src/components/text/text-utils.ts
@@ -1,7 +1,8 @@
 import { hasSpecificSlottedTag } from '../../utils';
+import { type TypographyAlign, type TypographyAlignDeprecated } from '../../types';
 
 export const TEXT_TAGS = ['p', 'span', 'div', 'address', 'blockquote', 'figcaption', 'cite', 'time', 'legend'] as const;
-export type TextTag = typeof TEXT_TAGS[number];
+export type TextTag = (typeof TEXT_TAGS)[number];
 
 export const getTextTagType = (host: HTMLElement, tag: TextTag): string => {
   if (hasSpecificSlottedTag(host, TEXT_TAGS.join())) {
@@ -10,3 +11,7 @@ export const getTextTagType = (host: HTMLElement, tag: TextTag): string => {
     return tag;
   }
 };
+
+/** @deprecated */
+export type TextAlignDeprecated = TypographyAlignDeprecated;
+export type TextAlign = TypographyAlign;